requirements.txt
azure-cognitiveservices-vision-customvision==3.1.1
azure-common==1.1.28
azure-core==1.34.0
azure-mgmt-core==1.5.0
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
fonttools==4.58.4
idna==3.10
isodate==0.7.2
kiwisolver==1.4.8
msrest==0.7.1
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pipreqs==0.4.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
yarg==0.1.10
